+++ src/Bridge/Milvus/Tests/IntegrationTest.php
@@ -1,0 +1,47 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\AI\Store\Bridge\Milvus\Tests;
+
+use PHPUnit\Framework\Attributes\Group;
+use Symfony\AI\Store\Bridge\Milvus\Store;
+use Symfony\AI\Store\StoreInterface;
+use Symfony\AI\Store\Test\AbstractStoreIntegrationTestCase;
+use Symfony\Component\HttpClient\HttpClient;
+
+/**
+ * @author Christopher Hertel <mail@christopher-hertel.de>
+ */
+#[Group('integration')]
+final class IntegrationTest extends AbstractStoreIntegrationTestCase
+{
+    protected static function createStore(): StoreInterface
+    {
+        return new Store(
+            HttpClient::create(),
+            'http://127.0.0.1:19530',
+            '',
+            'test_database',
+            'test_collection',
+            dimensions: 3,
+        );
+    }
+
+    protected static function getSetupOptions(): array
+    {
+        return ['forceDatabaseCreation' => true];
+    }
+
+    protected function waitForIndexing(): void
+    {
+        sleep(2);
+    }
+}