+++ tests/Double/TestStore.php
@@ -13,7 +13,7 @@ namespace Symfony\AI\Store\Tests\Double;
 
 use Symfony\AI\Platform\Vector\Vector;
 use Symfony\AI\Store\Document\VectorDocument;
-use Symfony\AI\Store\Exception\LogicException;
+use Symfony\AI\Store\Exception\UnsupportedFeatureException;
 use Symfony\AI\Store\StoreInterface;
 
 final class TestStore implements StoreInterface
@@ -37,7 +37,7 @@ final class TestStore implements StoreInterface
 
     public function remove(string|array $ids, array $options = []): void
     {
-        throw new LogicException('Method not implemented yet.');
+        throw new UnsupportedFeatureException('Method not implemented yet.');
     }
 
     public function query(Vector $vector, array $options = []): iterable